完成add推理任务表函数

+++ src/main/java/org/example/sentimentanalysis/model/InferenceTasks.java
@@ -66,7 +66,7 @@ public class InferenceTasks implements Serializable {
      * 评论数 (处理评论总数量)
      */
     @TableField("processed_count")
-    private Integer processedCount;
+    private Long processedCount;
 
     /**
      * 平均处理速度 (单位: 条/秒)
